Adding AtomicInt and CacheProtectedBytes
AtomicInt is a cacheline padded int64 value wrapped with various atomic
convenience methods.

CacheProtectedBytes will allocate a slice of bytes of a given size with
extra byte padding at each end to prevent false sharing. The caller only
sees a slice of the size they want and the padding is invisible.